Add vanilla JS MVC functions
Here we have our model, consisting of parts of the config file and server responses.

The view consists of functions that render the events and key events, setup the event listener on the key events and scroll function.

The handler consists of cuntions that initialise the app (model and view), fetch data from our local server and scroll.